requirements.txt
boto3==1.9.74
botocore==1.12.74
ffmpeg-python==0.1.17
sqlalchemy==1.2.17
psycopg2==2.7.7
requests==2.21.0